--- pkg/fuzzy/common/errors.go
@@ -1,9 +1,0 @@
-package common
-
-import "errors"
-
-var (
-	ErrNilMessages         error = errors.New("nil messages")
-	ErrNoMessages          error = errors.New("no messages")
-	ErrTargetLangUndefined error = errors.New("target language undefined")
-)
